Restructuring project #2

+++ Globus/Timeout/GLBTimeout.m
@@ -7,7 +7,7 @@
 @implementation GLBTimeout
 
 + (void)executeBlock:(GLBTimeoutBlock)block afterDelay:(NSTimeInterval)delay {
-    dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, delay * NSEC_PER_SEC);
+    dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(delay * NSEC_PER_SEC));
     dispatch_after(popTime, dispatch_get_main_queue(), block);
 }
